+++ packages/lib/storybook/stories/wallets/GooglePay.stories.tsx
@@ -19,192 +19,3 @@ const createComponent = (args, context) => {
 export const Default: Story = {
     render: createComponent
 };
-
-function getShippingCost(countryCode) {
-    switch (countryCode) {
-        case 'BR':
-            return {
-                'shipping-001': '5.00',
-                'shipping-002': '19.99'
-            };
-        default: {
-            return {
-                'shipping-001': '0.00',
-                'shipping-002': '1.99',
-                'shipping-003': '9.99'
-            };
-        }
-    }
-}
-
-function getShippingOptions(countryCode?: string) {
-    switch (countryCode) {
-        case 'BR': {
-            return {
-                defaultSelectedOptionId: 'shipping-001',
-                shippingOptions: [
-                    {
-                        id: 'shipping-001',
-                        label: '$5.00: Standard shipping',
-                        description: 'Free Shipping delivered in 10 business days.'
-                    },
-                    {
-                        id: 'shipping-002',
-                        label: '$19.99: Standard shipping',
-                        description: 'Standard shipping delivered in 2 business days.'
-                    }
-                ]
-            };
-        }
-        default: {
-            return {
-                defaultSelectedOptionId: 'shipping-001',
-                shippingOptions: [
-                    {
-                        id: 'shipping-001',
-                        label: '$0.00: Free shipping',
-                        description: 'Free Shipping delivered in 10 business days.'
-                    },
-                    {
-                        id: 'shipping-002',
-                        label: '$1.99: Standard shipping',
-                        description: 'Standard shipping delivered in 3 business days.'
-                    },
-                    {
-                        id: 'shipping-003',
-                        label: '$9.99: Express shipping',
-                        description: 'Express shipping delivered in 1 business day.'
-                    }
-                ]
-            };
-        }
-    }
-}
-
-function getTransactionInfo(): google.payments.api.TransactionInfo {
-    return {
-        displayItems: [
-            {
-                label: 'Subtotal',
-                type: 'SUBTOTAL',
-                price: '99.00'
-            },
-            {
-                label: 'Tax',
-                type: 'TAX',
-                price: '1.00'
-            }
-        ],
-        countryCode: 'US',
-        currencyCode: 'USD',
-        totalPriceStatus: 'FINAL',
-        totalPrice: '12.00',
-        totalPriceLabel: 'Total'
-    };
-}
-
-function calculateNewTransactionInfo(countryCode: string, selectedShippingOptionId: string) {
-    const newTransactionInfo = getTransactionInfo();
-    const shippingCost = getShippingCost(countryCode)[selectedShippingOptionId];
-
-    newTransactionInfo.displayItems.push({
-        type: 'LINE_ITEM',
-        label: 'Shipping cost',
-        price: shippingCost,
-        status: 'FINAL'
-    });
-
-    let totalPrice = 0.0;
-    newTransactionInfo.displayItems.forEach(displayItem => (totalPrice += parseFloat(displayItem.price)));
-    newTransactionInfo.totalPrice = totalPrice.toString();
-
-    return newTransactionInfo;
-}
-
-/**
- * TODO:
- * - disable sessions for this flow
- * - get shopper details and expose them
- * - test if we can reject the payment with the google overlay open
- */
-export const Express: Story = {
-    render: createComponent,
-    argTypes: {
-        countryCode: {
-            control: false
-        },
-        amount: {
-            control: false
-        }
-    },
-    args: {
-        countryCode: 'US',
-        amount: 10000,
-        // TODO: Check difference of onAuthorized vs onPaymentAuthorized -
-        componentConfiguration: {
-            // Fields 'countryCode,currencyCode,totalPriceStatus,totalPrice' are populated on requests.ts
-            transactionInfo: getTransactionInfo(),
-            callbackIntents: ['SHIPPING_ADDRESS', 'SHIPPING_OPTION', 'PAYMENT_AUTHORIZATION'],
-            paymentDataCallbacks: {
-                onPaymentDataChanged(intermediatePaymentData) {
-                    return new Promise(resolve => {
-                        const { callbackTrigger, shippingAddress, shippingOptionData } = intermediatePaymentData;
-                        const paymentDataRequestUpdate: google.payments.api.PaymentDataRequestUpdate = {};
-
-                        /** Validate country/address selection  **/
-                        if (shippingAddress.countryCode !== 'US' && shippingAddress.countryCode !== 'BR') {
-                            paymentDataRequestUpdate.error = {
-                                reason: 'SHIPPING_ADDRESS_UNSERVICEABLE',
-                                message: 'Cannot ship to the selected address',
-                                intent: 'SHIPPING_ADDRESS'
-                            };
-                        }
-                        /** If it initializes or changes the shipping address, we calculate the shipping options and transaction info  */
-                        if (callbackTrigger === 'INITIALIZE' || callbackTrigger === 'SHIPPING_ADDRESS') {
-                            paymentDataRequestUpdate.newShippingOptionParameters = getShippingOptions(shippingAddress.countryCode);
-                            const selectedShippingOptionId = paymentDataRequestUpdate.newShippingOptionParameters.defaultSelectedOptionId;
-                            paymentDataRequestUpdate.newTransactionInfo = calculateNewTransactionInfo(
-                                shippingAddress.countryCode,
-                                selectedShippingOptionId
-                            );
-                        }
-                        /** If SHIPPING_OPTION changed, we calculate the new fee */
-                        if (callbackTrigger === 'SHIPPING_OPTION') {
-                            paymentDataRequestUpdate.newTransactionInfo = calculateNewTransactionInfo(
-                                shippingAddress.countryCode,
-                                shippingOptionData.id
-                            );
-                        }
-
-                        resolve(paymentDataRequestUpdate);
-                    });
-                },
-                /**
-                 * Shouldn't we be always using this?
-                 * https://developers.google.com/pay/api/web/guides/tutorial#handle-onpaymentauthorized-callbacacks
-                 *
-                 * @param paymentData
-                 */
-                onPaymentAuthorized(paymentData) {
-                    console.log('onPaymentAuthorized', paymentData);
-
-                    return Promise.resolve({
-                        transactionState: 'ERROR',
-                        error: {
-                            intent: 'PAYMENT_AUTHORIZATION',
-                            message: 'Insufficient funds',
-                            reason: 'PAYMENT_DATA_INVALID'
-                        }
-                    });
-                }
-            },
-            // Shipping Address config
-            shippingAddressRequired: true,
-            shippingAddressParameters: {
-                phoneNumberRequired: true
-            },
-            // Shipping Options config
-            shippingOptionRequired: true
-        }
-    }
-};
